feat: Add read_vector function for parsing strings into vectors
This commit adds a new function called read_vector to the pipe namespace. This function takes a string as input and parses it into a vector of the specified type. It uses std::istringstream and std::copy to achieve this.

The read_vector function is used in the pipe_test.cc file to test the functionality of the pipe library. It is used to parse strings into vectors of different types, such as strings, integers, doubles, and characters. The parsed vectors are then compared with the expected vectors using the REQUIRE macro.

This addition improves the flexibility and usability of the pipe library by providing a convenient way to parse strings into vectors.