feat: Add FNV1a hash implementation
This commit adds a new file `fnv1a.hh` to the `sugar` namespace. It contains an implementation of the FNV1a hash algorithm, specifically the `fnv1a_32` function. This function takes a string as input and calculates the 32-bit FNV1a hash value.

The addition of the FNV1a hash implementation enhances the functionality of the `sugar` namespace by providing a reliable and efficient hash algorithm for string hashing.