feat: Add factory.hh and factory_test.cc
The code changes involve adding a new file `factory.hh` and `factory_test.cc` to the `sugar` namespace. The `factory.hh` file contains the implementation of a factory template class for getting and registering product objects. The `factory_test.cc` file contains unit tests for the factory functionality.

This addition enhances the functionality of the `sugar` namespace by providing a flexible and efficient way to create and manage product objects using the factory pattern.