feat: Add join function to the pipe namespace
This commit adds a new function called join to the pipe namespace. This function takes an optional separator string as input and joins the elements of a vector into a single string. If a separator is provided, it is inserted between each pair of elements. The joined string is then returned.

The join function is used in the pipe_test.cc file to test the functionality of the pipe library. It is used to join strings in a vector into a single string with different separators. The joined strings are then compared with the expected results using the REQUIRE macro.

This addition enhances the functionality of the pipe library by providing a convenient way to join elements of a vector into a single string.